+++ CHANGELOG.md
@@ -7,7 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
-- Sort loaded logs alphabetically
+## [1.24.0]
+
+- Filter invalid values from the `frame-altimeter` into a "invalid values" dataset, just like with `navsys`
+- Group filter plot labels after log type
+- Group loaded logs after type
+- Sort loaded logs alphabetically and after start time
 - Show loaded logs start datetime as `YYYY-MM-DD HH:MM:SS`
 
 ## [1.23.0]